requirements-dev.txt
-r requirements.txt
ruff==0.3.2
isort==5.13.2
flake8==7.0.0
pylint==3.1.0
mypy==1.9.0
pyright==1.1.353
pandas-stubs==2.2.0.240218
